Revised to prepare for other platforms (linux initially)

+++ components/comms/CommsChannels/CommsChannelManager.cpp
@@ -332,7 +332,7 @@ void CommsChannelManager::inboundHandleMsgVec(uint32_t channelID, const SpiramAw
     // Check the channel
     if (channelID >= _commsChannelVec.size())
     {
-        LOG_W(MODULE_PREFIX, "inboundHandleMsg, channelID channelId %d is INVALID msglen %d", channelID, msg.size());
+        LOG_W(MODULE_PREFIX, "inboundHandleMsg, channelID channelId %d is INVALID msglen %d", channelID, (int)msg.size());
         return;
     }
 
@@ -340,7 +340,7 @@ void CommsChannelManager::inboundHandleMsgVec(uint32_t channelID, const SpiramAw
     CommsChannel* pChannel = _commsChannelVec[channelID];
     if (!pChannel)
     {
-        LOG_W(MODULE_PREFIX, "inboundHandleMsg, channelID channelId %d is NULL msglen %d", channelID, msg.size());
+        LOG_W(MODULE_PREFIX, "inboundHandleMsg, channelID channelId %d is NULL msglen %d", channelID, (int)msg.size());
         return;
     }
 
@@ -584,7 +584,7 @@ void CommsChannelManager::ensureProtocolCodecExists(uint32_t channelID)
     }
 
     // Debug
-    LOG_W(MODULE_PREFIX, "No suitable codec found for protocol %s map entries %d", channelProtocol.c_str(), _protocolCodecFactoryList.size());
+    LOG_W(MODULE_PREFIX, "No suitable codec found for protocol %s map entries %d", channelProtocol.c_str(), (int)_protocolCodecFactoryList.size());
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////
